refactor to use worker_base for both cuda and tpu workers
